geo/geos: compile geo/geos package on Windows
This commits adds support for pkg/geo/geos to be compiled on Windows
platforms without separating by OS. This stops the split from
geos_unix.go and geos_windows.go (all back into geos.go). We've done
some #define magic for dlopen/dlsym/dlclose/dlerror to use the
equivalent windows functions.

This should stop any pesky skews between the Windows and Unix build.
Note that GEOS is not compiled in the Makefile, so this isn't tested.
What's important for now is that everything builds without skew in
a path that will commonly change.

Release note: None
