+++ pkg/geo/geos/geos_test.go
@@ -8,8 +8,6 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-// +build !windows
-
 package geos
 
 import (
